refactor: move utility functions to separate file
